fix(career): keep zh workbook repair outside runtime services
