+++ docs/codex/pr-train-state.json
@@ -52768,7 +52768,7 @@
     "repo": "fap-api",
     "branch": "codex/career-zh-parity-fix-02",
     "base": "origin/main",
-    "status": "github_checks_pending",
+    "status": "github_checks_pending_after_scoped_fix",
     "train_scope": "career_zh_parity_fix",
     "title": "CAREER-ZH-PARITY-FIX-02: fix(career): repair reviewed zh display workbook contract",
     "depends_on": [
@@ -52785,11 +52785,11 @@
       },
       "implementation": {
         "status": "pass",
-        "evidence": "Added career:repair-display-workbook-contract and CareerDisplayWorkbookContractRepairer. Command defaults to dry-run, writes no DB/CMS data, and only writes a repaired XLSX when --execute and --output are provided."
+        "evidence": "Added career:repair-display-workbook-contract with embedded workbook contract repair logic. Command defaults to dry-run, writes no DB/CMS data, and only writes a repaired XLSX when --execute and --output are provided."
       },
       "php_lint": {
         "status": "pass",
-        "evidence": "php -l passed for CareerRepairDisplayWorkbookContract.php, CareerDisplayWorkbookContractRepairer.php, and CareerRepairDisplayWorkbookContractCommandTest.php."
+        "evidence": "php -l passed for CareerRepairDisplayWorkbookContract.php after moving repair logic into the command to avoid runtime service diff."
       },
       "focused_tests": {
         "status": "pass",
@@ -52809,7 +52809,7 @@
       },
       "format": {
         "status": "pass",
-        "evidence": "cd backend && ./vendor/bin/pint --test app/Console/Commands/CareerRepairDisplayWorkbookContract.php app/Services/Career/Import/CareerDisplayWorkbookContractRepairer.php tests/Feature/Console/CareerRepairDisplayWorkbookContractCommandTest.php passed."
+        "evidence": "cd backend && ./vendor/bin/pint --test app/Console/Commands/CareerRepairDisplayWorkbookContract.php tests/Feature/Console/CareerRepairDisplayWorkbookContractCommandTest.php passed after removing the runtime service file from the final PR diff."
       },
       "metadata": {
         "status": "pass",
@@ -52817,7 +52817,7 @@
       },
       "scope_validation": {
         "status": "pass",
-        "evidence": "Changed files are confined to PR02 scoped command, import service, focused test, generated career report, and docs/codex metadata."
+        "evidence": "Changed files are confined to PR02 scoped command, focused test, generated career report, and docs/codex metadata; backend/app/Services/Career/Import/CareerDisplayWorkbookContractRepairer.php was removed from final PR diff after GitHub verify-bigfive scope-freeze failure."
       },
       "diff_check": {
         "status": "pass",
@@ -52826,10 +52826,14 @@
       "cached_diff_check": {
         "status": "pass",
         "evidence": "git diff --cached --check passed after path-limited staging of PR02 scoped files."
+      },
+      "github_verify_bigfive_initial": {
+        "status": "failed_fixed_in_scope",
+        "evidence": "GitHub verify-bigfive initially failed because BigFiveResultPageV2CoreBodyPreviewTest::test_runtime_paths_have_no_uncommitted_diff detected backend/app/Services/Career/Import/CareerDisplayWorkbookContractRepairer.php. The fix moved repair logic into the command and removed that runtime service path from the final PR diff."
       }
     },
     "failure_reason": null,
-    "commit_sha": "112100e088c1d53b847cf87552a0de125b6fc0ad",
+    "commit_sha": "405af96e97921c2369c842d1410a2c3cae8e44d9",
     "pr_url": "https://github.com/fermatmind/fap-api/pull/2026",
     "merged_at": null,
     "remote_branch_deleted": false,
@@ -52863,6 +52867,16 @@
         "at": "2026-06-11T03:28:00Z",
         "status": "github_checks_pending",
         "note": "Opened PR #2026 for branch codex/career-zh-parity-fix-02 and pushed local verification metadata; waiting for GitHub checks."
+      },
+      {
+        "at": "2026-06-11T04:05:00Z",
+        "status": "github_checks_pending_after_scoped_fix",
+        "note": "GitHub verify-bigfive failed on the runtime service path; scoped fix moved repair logic into the command, removed the runtime service from final PR diff, and re-ran php lint, focused feature test, real workbook dry-run, execute+second-pass probe, and scoped Pint locally."
+      },
+      {
+        "at": "2026-06-11T04:08:00Z",
+        "status": "committed",
+        "note": "Committed scoped CI fix as 405af96e97921c2369c842d1410a2c3cae8e44d9; final origin/main...HEAD diff no longer includes backend/app/Services/Career/Import/**."
       }
     ]
   },
